Fix editing & new injections
